refactor: LogBatFactory 로직 변경
- 싱글턴 보장을 위한 VarHandler 사용 로직을 적용했습니다.
- javadoc을 추가했습니다.
